requirements.txt
pymysql==1.0.2
whitenoise==5.3.0
schedule==1.1.0
Django==3.2.*
djangorestframework==3.12.*
djangorestframework-simplejwt==4.7.*
django-mysql==3.10.0
mysql-connector-python==8.0.25
dj-database-url==0.5.0
django-cors-headers==3.7.*
Pillow==8.4.0
psycopg2